+++ include/zeus/crypto_service.h
@@ -11,6 +11,7 @@ enum class DigestAlgorithm {
     Md5,
     Sha1,
     Sha256,
+    Sha384,
     Sha512,
     Crc32,
 };